+++ filepanther/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 
 from filepanther.util.config_logger import config_logger
 config_logger()