+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
 * allows parsing only basenames
+* adds special metadata extraction of `_datetime`
 
 ## 1.0.0
 * Rethought organization that simplifies API but eliminates db connections.